+++ plugin/commands/theme/component/build.js
@@ -1,22 +1,27 @@
 import { Args, Command, Flags } from '@oclif/core'
-import { spawn } from 'node:child_process'
+import { decodeToml } from '../../../../src/utils/TomlUtils.js'
+import { readFileSync } from 'node:fs'
+import { CONFIG_FILE_NAME } from '../../../../src/config/CLI.js'
+import SessionFactory from '../../../../src/factory/SessionFactory.js'
+import BuildCommand from '../../../../src/commands/BuildCommand.js'
 
 export default class Build extends Command {
+  static description = 'Build a component or collection of components'
+
   static args = {
     component: Args.string({ description: 'Component to build' })
   }
 
-  static description = 'Build a component or collection of components'
-
   static flags = {
-    watch: Flags.boolean({ char: 'w', description: 'Watch for changes' })
+    watch: Flags.boolean({ char: 'w', description: 'Watch for changes' }),
+    debug: Flags.boolean({ description: 'Debug Mode is more verbose.' }),
+    trace: Flags.boolean({ description: 'Trace Mode provides tracing and debug information.' })
   }
 
   async run () {
     const { args, flags } = await this.parse(Build)
 
-    const command = 'npm'
-    const commandArgs = ['exec', '--', 'component', 'build']
+    const commandArgs = ['build']
     if (args.component) {
       commandArgs.push(args.component)
     }
@@ -25,15 +30,9 @@ export default class Build extends Command {
       commandArgs.push('--watch')
     }
 
-    const componentProcess = spawn(command, commandArgs, {
-      stdio: 'inherit'
-    })
-
     // Wait for component process to complete
-    await new Promise((resolve) => {
-      componentProcess.on('close', code => {
-        resolve(code)
-      })
-    })
+    const config = /** @type {{component: import('../models/static/Session.js').CLIConfig}} */ (decodeToml(readFileSync(CONFIG_FILE_NAME, 'utf8')))
+    SessionFactory.fromArgsAndManifest(commandArgs, config)
+    await BuildCommand.execute()
   }
 }