+++ plugin/commands/theme/component/generate.js
@@ -1,29 +1,30 @@
-import { Args, Command } from '@oclif/core'
-import { spawn } from 'node:child_process'
+import { Args, Command, Flags } from '@oclif/core'
+import { decodeToml } from '../../../../src/utils/TomlUtils.js'
+import { readFileSync } from 'node:fs'
+import { CONFIG_FILE_NAME } from '../../../../src/config/CLI.js'
+import SessionFactory from '../../../../src/factory/SessionFactory.js'
+import GenerateCommand from '../../../../src/commands/GenerateCommand.js'
 
 export default class Generate extends Command {
+  static description = 'Generate a component'
+
   static args = {
     component: Args.string({ description: 'Component to generate', required: true })
   }
 
-  static description = 'Generate a component'
+  static flags = {
+    debug: Flags.boolean({ description: 'Debug Mode is more verbose.' }),
+    trace: Flags.boolean({ description: 'Trace Mode provides tracing and debug information.' })
+  }
 
   async run () {
     const { args } = await this.parse(Generate)
 
-    const command = 'npm'
-
-    const commandArgs = ['exec', 'component', 'generate', 'component', args.component]
-
-    const componentProcess = spawn(command, commandArgs, {
-      stdio: 'inherit'
-    })
+    const commandArgs = ['generate', 'component', args.component]
 
     // Wait for component process to complete
-    await new Promise((resolve) => {
-      componentProcess.on('close', code => {
-        resolve(code)
-      })
-    })
+    const config = /** @type {{component: import('../models/static/Session.js').CLIConfig}} */ (decodeToml(readFileSync(CONFIG_FILE_NAME, 'utf8')))
+    SessionFactory.fromArgsAndManifest(commandArgs, config)
+    await GenerateCommand.execute(config)
   }
 }